+++ src/main/java/at/hannibal2/skyhanni/events/BlazeParticleEvent.kt
@@ -1,7 +1,0 @@
-package at.hannibal2.skyhanni.events
-
-import net.minecraft.entity.monster.EntityBlaze
-import net.minecraftforge.fml.common.eventhandler.Cancelable
-
-@Cancelable
-class BlazeParticleEvent(val blaze: EntityBlaze): LorenzEvent()