+++ src/main/java/at/hannibal2/skyhanni/utils/LorenzUtils.kt
@@ -113,4 +113,19 @@ object LorenzUtils {
     fun consoleLog(text: String) {
         SkyHanniMod.consoleLog(text)
     }
+
+    // Taken and modified from https://stackoverflow.com/a/11306854/5507634
+    fun getCallerClass(vararg skip: String): String? {
+        val stElements = Thread.currentThread().stackTrace
+        for (i in 1 until stElements.size) {
+            val ste = stElements[i]
+            val className = ste.className
+            if (className != LorenzUtils::class.java.name && className.indexOf("java.lang.Thread") != 0) {
+                if (className !in skip) {
+                    return className
+                }
+            }
+        }
+        return null
+    }
 }